Move out reducer to separate file
